\name{cooks.distance}
\alias{cooks.distance.lad}
\alias{cooks.distance.nls}
\alias{cooks.distance.ols}
\alias{cooks.distance.ridge}
\title{Cook's distances}
\usage{
\method{cooks.distance}{ols}(model, \dots)
\method{cooks.distance}{nls}(model, \dots)
\method{cooks.distance}{lad}(model, type = "L2", \dots)
\method{cooks.distance}{ridge}(model, type = "cov", \dots)
}
\arguments{
  \item{model}{ an \R object, returned by \code{ols}, \code{\link{nls}}, \code{lad} or \code{ridge}.}
  \item{type}{ For \code{'lad'} objects, specifies the norm used for the influence measure. \code{"L2"} 
  computes the Cook's distance based on the Euclidean norm, whereas \code{"L1"} computes the absolute-norm
  fitted-value measure \eqn{FD_{1i}} proposed in Equation (4) of Sun and Wei (2004).

  For \code{'ridge'} objects, the available options are \code{"1st"}, \code{"cov"} and \code{"both"}, 
  corresponding to the Cook's distances given by Equations (2.5), (2.6), or both, respectively, in
  Walker and Birch (1988).}
  \item{\dots}{ further arguments passed to or from other methods.}
}
\description{
  Cook's distance is a measure to assess the influence of the \emph{i}th observation on the 
  model parameter estimates. This function computes the Cook's distance based on leave-one-out 
  cases deletion for ordinary least squares, nonlinear least squares, lad and ridge regression.
}
\value{
  A vector whose \emph{i}th element contains the Cook's distance,
  \deqn{D_i(\bold{M},c) = \frac{(\hat{\bold{\beta}}_{(i)} - \hat{\bold{\beta}})^T\bold{M}
  (\hat{\bold{\beta}}_{(i)} - \hat{\bold{\beta}})}{c},}
  for \eqn{i = 1,\dots,n}, with \eqn{\bold{M}} a positive definite matrix and \eqn{c > 0}. Specific 
  choices of \eqn{\bold{M}} and \eqn{c} are done for objects of class \code{ols}, \code{\link{nls}}, 
  \code{lad} and \code{ridge}.

  The Cook's distance for nonlinear regression is based on linear approximation, which may be 
  inappropriate for expectation surfaces markedly nonplanar.

  For \code{lad} objects with \code{type = "L1"}, the returned measure is
  \deqn{
    FD_i = \frac{1}{\hat{\omega}}\big\|\bold{X}\hat{\boldsymbol{\beta}} 
    - \bold{X}\hat{\boldsymbol{\beta}}_{(i)}\big\|_1 = \frac{1}{\hat{\omega}}
    \sum_{j=1}^n\big|\bold{x}_j^T\hat{\boldsymbol{\beta}} - \bold{x}_j^T
    \hat{\boldsymbol{\beta}}_{(i)}\big|,
  }
  where
  \deqn{
    \hat{\omega} = \frac{1}{n-p}\sum_{j=1}^n|y_j - \bold{x}_j^T\hat{\boldsymbol{\beta}}|.
  }
}
\references{
  Cook, R.D., Weisberg, S. (1980).
  Characterizations of an empirical influence function for detecting influential cases in regression.
  \emph{Technometrics} \bold{22}, 495-508.
  \doi{10.1080/00401706.1980.10486199}

  Cook, R.D., Weisberg, S. (1982).
  \emph{Residuals and Influence in Regression}.
  Chapman and Hall, London.

  Ross, W.H. (1987).
  The geometry of case deletion and the assessment of influence in nonlinear regression.
  \emph{The Canadian Journal of Statistics} \bold{15}, 91-103.
  \doi{10.2307/3315198}

  Sun, R.B., Wei, B.C. (2004).
  On influence assessment for LAD regression.
  \emph{Statistics & Probability Letters} \bold{67}, 97-110.
  \doi{10.1016/j.spl.2003.08.018}

  Walker, E., Birch, J.B. (1988).
  Influence measures in ridge regression.
  \emph{Technometrics} \bold{30}, 221-227.
  \doi{10.1080/00401706.1988.10488370}
}
\examples{
# Cook's distances for linear regression
fm <- ols(stack.loss ~ ., data = stackloss)
CD <- cooks.distance(fm)
plot(CD, ylab = "Cook's distances", ylim = c(0,0.8))
text(21, CD[21], label = as.character(21), pos = 3)

# Cook's distances for LAD regression
fm <- lad(stack.loss ~ ., data = stackloss)
CD <- cooks.distance(fm, type = "L2")
plot(CD, ylab = "Cook's distances", ylim = c(0,0.4))
text(17, CD[17], label = as.character(17), pos = 3)

# Fitted-value measure, L1-norm for LAD regression
FD <- cooks.distance(fm, type = "L1")
plot(FD, ylab = "Fitted-value distances", ylim = c(0,2.3))
text(17, FD[17], label = as.character(17), pos = 3)

# Cook's distances for ridge regression
data(portland)
fm <- ridge(y ~ ., data = portland)
CD <- cooks.distance(fm)
plot(CD, ylab = "Cook's distances", ylim = c(0,0.5))
text(8, CD[8], label = as.character(8), pos = 3)

# Cook's distances for nonlinear regression
data(skeena)
model <- recruits ~ b1 * spawners * exp(-b2 * spawners)
fm <- nls(model, data = skeena, start = list(b1 = 3, b2 = 0))
CD <- cooks.distance(fm)
plot(CD, ylab = "Cook's distances", ylim = c(0,0.35))
obs <- c(5, 6, 9, 19, 25)
text(obs, CD[obs], label = as.character(obs), pos = 3)
}
\keyword{regression}
