\name{portland}
\docType{data}
\alias{portland}

\title{Portland cement dataset}
\description{
  This dataset comes from an experimental investigation of the heat evolved during the 
  setting and hardening of Portland cements of varied composition and the dependence of 
  this heat on the percentages of four compounds in the clinkers from which the cement 
  was produced.
}
\usage{data(portland)}
\format{
  A data frame with 13 observations on the following 5 variables.
  \describe{
    \item{y}{ The heat evolved after 180 days of curing, measured in calories per gram of cement.}
    \item{x1}{ Tricalcium aluminate.}
    \item{x2}{ Tricalcium silicate.}
    \item{x3}{ Tetracalcium aluminoferrite.}
    \item{x4}{ \eqn{\beta}-dicalcium silicate.}
  }
}
\source{
  Kaciranlar, S., Sakallioglu, S., Akdeniz, F., Styan, G.P.H., Werner, H.J. (1999).
  A new biased estimator in linear regression and a detailed analysis of the widely-analysed dataset on Portland cement.
  \emph{Sankhya, Series B} \bold{61}, 443-459.
}
\keyword{datasets}