\name{skeena}
\docType{data}
\alias{skeena}

\title{Skeena River sockeye salmon data}
\description{
  The data have 28 observations of spawners and recruits (units are thousands of fish) 
  from 1940 until 1967 for the Skeena river sockeye salmon stock.
}
\usage{data(skeena)}
\format{
  A data frame with 28 observations on the following 3 variables.
  \describe{
    \item{year}{ Years in which the number of spawners and recruits were recorded.}
    \item{spawners}{ Size of the annual spawning stock.}
    \item{recruits}{ Production of new catchable-sized fish.}
  }
}
\source{
  Carroll, R.J., Ruppert, D. (1988).
  \emph{Transformation and Weighting in Regression}.
  Chapman and Hall, London.
}
\keyword{datasets}